Add to shouldFeatureRun
