+++ source/feature-manager.tsx
@@ -9,8 +9,12 @@ import waitFor from './helpers/wait-for.js';
 import onAbort from './helpers/abort-controller.js';
 import ArrayMap from './helpers/map-of-arrays.js';
 import bisectFeatures from './helpers/bisect.js';
-import {shouldFeatureRun, PromisableBooleanFunction, BooleanFunction} from './github-helpers/index.js';
-import {isFeaturePrivate} from './helpers/feature-utils.js';
+import {
+	PromisableBooleanFunction,
+	BooleanFunction,
+	shouldFeatureRun,
+	isFeaturePrivate,
+} from './helpers/feature-utils.js';
 import optionsStorage, {isFeatureDisabled, RGHOptions} from './options-storage.js';
 import {
 	applyStyleHotfixes,
